+++ internal/music/music.go
@@ -163,9 +163,8 @@ func (m *Music) TrackURL(t Track) *url.URL {
 }
 
 // Find track using the etag from the S3 bucket.
-func (m *Music) TrackLookup(etag string) Track {
-	track, _ := m.LookupETag(etag)
-	return track
+func (m *Music) TrackLookup(etag string) (Track, error) {
+	return m.LookupETag(etag)
 }
 
 // URL for track cover image.